requirements.txt
requests==2.31.0
pandas==2.0.1
python-dotenv==1.0.0
pyarrow==12.0.1
fastparquet==2023.4.0
openai==0.27.6
tqdm==4.65.0
retrying==1.3.4
numpy==1.24.3
psutil
boto3==1.28.38
botocore==1.31.38
# Data analysis and visualization
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
# Optional: for interactive visualizations in notebooks
plotly==5.14.1
ipywidgets==8.0.6